fix all kinds of stupidity from me (DShadowWolf) and start moving this ahead towards 'ready to merge' status

+++ src/main/java/com/mcmoddev/lib/integration/plugins/thaumcraft/IAspectCalculation.java
@@ -1,7 +1,5 @@
 package com.mcmoddev.lib.integration.plugins.thaumcraft;
 
-import thaumcraft.api.aspects.Aspect;
-
 /**
  * Allowance for a Lambda/anon-method for calculating the amount of an aspect
  * per item and/or block.
@@ -10,6 +8,5 @@ import thaumcraft.api.aspects.Aspect;
  * @since 26-OCT-2018
  */
 public interface IAspectCalculation {
-	public float apply(float mult);
-//	public boolean canApply(TCMaterial material, Aspect aspect);
+	public int apply(float mult);
 }
